# Backend/OpenAPI.yaml
openapi: 3.0.3
info:
  title: AnaECommerce Backend API
  version: v1
  description: Backend REST API for AnaECommerce with JWT authentication and role-based access.
servers:
  - url: http://localhost:5001/api
    description: Development
  - url: https://ana.runasp.net/api
    description: Production
paths:
...
  /api/users:
    get:
      summary: Get users (admin only)
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUser'
      security:
        - BearerAuth: []
      tags:
        - Users
  /api/users/{id}:
    get:
      summary: Get a user by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200': {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserDto"}}}}
        '404': {"description": "Not Found"}
      security:
        - BearerAuth: []
      tags:
        - Users
    put:
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200': {"description": "OK"}
        '400': {"description": "Bad Request"}
        '404': {"description": "Not Found"}
      security:
        - BearerAuth: []
      tags:
        - Users
    delete:
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204': {"description": "No Content"}
        '404': {"description": "Not Found"}
      security:
        - BearerAuth: []
      tags:
        - Users
  /api/users/{id}/role:
    put:
      summary: Assign role to user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleDto'
      responses:
        '200': {"description": "OK"}
        '400': {"description": "Bad Request"}
      security:
        - BearerAuth: []
      tags:
        - Users
  /api/profile/GetProfile:
    get:
      summary: Get current user profile
      responses:
        '200': {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserProfileDto"}}}}
      security:
        - BearerAuth: []
      tags:
        - Profile
  /api/profile/UpdateProfile:
    put:
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileDto'
      responses:
        '200': {"description": "OK"}
      security:
        - BearerAuth: []
      tags:
        - Profile
  /api/profile/addresses:
    post:
      summary: Add new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressDto'
      responses:
        '201': {"description": "Created"}
      security:
        - BearerAuth: []
      tags:
        - Profile
  /api/profile/addresses/{id}:
    delete:
      summary: Delete address
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204': {"description": "No Content"}
      security:
        - BearerAuth: []
      tags:
        - Profile
  /api/profile/GetPaymentMethods:
    get:
      summary: Get payment methods
      responses:
        '200': {"description": "OK"}
      security:
        - BearerAuth: []
      tags:
        - Profile
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
    PaginatedUser:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        totalCount:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
    UpdateUserDto:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
    UserProfileDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    AddressDto:
      type: object
      properties:
        id:
          type: string
        line1:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
...